int -> uint16_t

+++ bsengine/src/bstorm/stage_types.hpp
@@ -2,5 +2,5 @@
 
 namespace bstorm
 {
-using StageIndex = int;
+using StageIndex = uint16_t;
 }
